requirements.txt
colorama==0.4.6
cx-Freeze==6.15.5
imgviz==1.6.2
labelme==5.1.1
loguru==0.6.0
matplotlib==3.2.2
natsort==8.2.0
numpy==1.24.1
opencv-python==4.7.0.68
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
PyYAML==6.0
QtPy==2.3.0
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
# Editable install with no version control (setuptools==65.6.3)
-e c:\users\hp\miniconda3\envs\labelme\lib\site-packages\setuptools-65.6.3-py3.9.egg-info
tensorflow-intel==2.11.0
termcolor==2.2.0
tifffile==2023.3.15
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1

WARNING(pigar): some manual fixes are required since pigar has found duplicate requirements for the same import name.
# WARNING(pigar): the following duplicate requirements are for import name: PyQt5
PyQt5==5.15.4
PyQt5-sip==12.11.0
